day 10 2023 done finally
After a lot of thinking..

Just move along the loop and add all points on one side to an inner loop
`c`. Then iteratively add neighbors of `c` to `c` (that are not on the
loop).

Just need to make sure that we add points on the correct side of the
loop.. :)